feat!: change NewClient to accept *http.Client
This commit changes the NewClient function to accept a pointer to an
http.Client (*http.Client) instead of a value (http.Client).  This is a breaking change for any code outside this package that uses NewClient.

The change was made to enable connection pooling and improve performance. Users of NewClient must now pass a pointer to an http.Client instance.

Example of the required change:

Before:
  client := NewClient(http.Client{}, mss.Credentials{}) // This will no longer compile

After:
  client := NewClient(&http.Client{}, mss.Credentials{})